requirements.txt
Django==1.3.1
django-appconf==0.4.1
django-autoslug==1.4.1
django_compressor==1.1.1
django-debug-toolbar==0.8.5
django-haystack==1.2.5
django-mptt==0.5.0
django-social-auth==0.6.0
Fabric==1.3.3
httplib2==0.7.1
Markdown==2.1.0
oauth2==1.5.170
psycopg2==2.4.2
pycrypto==2.4.1
python-openid==2.2.5
recaptcha-client==1.0.6
South==0.7.3
ssh==1.7.11
